feature: welcome header for cli introduction
As about the UX of nodeJS CLI we make a header with a welcome screen with what is being run, the description of the CLI, the number or version and if there is a tagline. In order to use  the fields of  package.json  inside  index.js  to create sort of a welcome screen. We go and add  cli-welcome package and remove  clear-any-console  package because it is more powerful.