add pack: cli-handle-unhandled that handles the unhandled errors
As about Error Handling, in order to handle the unhandled errors, as rejected promise, if we not handled this particular promise, then CLI is going to break. The Error message is  UnhandledPromiseRejectionWarning  For that, we can use  process.on()  in which we pass the event and then callback. This callback is going to go ahead and do something like console logging.
process.on('unhandledRejection', (err) => console.log('unhandledRejection', err))
Promise.reject(new Error('THIS_IS_UNHANDLED'))
We are going to use  cli-handle-unhandled  package for help in this situation.