Node.js process.argv arguments and parsing CLI arguments
Node process.argv returns an array containing the command line arguments passed to a nodejs process
On executing  console.log(process.argv)  inside   index.js  gives the nodejs executable path, also the path the file that is running this process, the given flag and the input for this flag. Args starts from the third item of array.